astro_backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# AI & Astrology
google-genai>=0.3.0
kerykeion>=4.0.0

# Utils
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
